improved post functionality

+++ routes/posts.js
@@ -1,29 +1,37 @@
 const router = require("express").Router();
 
 const Post = require("../models/Post");
+const getUser = require("../middleware/getUser");
+const auth = require("../middleware/auth");
 
 //get all the posts
-router.get("/", async (req, res) => {
+router.get("/", getUser, async (req, res) => {
   try {
     const posts = await Post.find();
     res.send(posts);
   } catch (err) {
     console.log(error);
+    res
+      .status(500)
+      .status({ message: "Something went wrong. Please try again" });
   }
 });
 
 //get a single post
-router.get("/:postId", async (req, res) => {
+router.get("/:postId", getUser, async (req, res) => {
   try {
     const post = await Post.findById(req.params.postId);
-    res.send(post);
+    res.status(200).send({ post });
   } catch (err) {
     console.log(err);
+    res
+      .status(500)
+      .status({ message: "Something went wrong. Please try again" });
   }
 });
 
 //Create a new post
-router.post("/", async (req, res) => {
+router.post("/", getUser, auth, async (req, res) => {
   const { userName, gameTitle, genre, description } = req.body;
   const post = new Post({
     userName,
@@ -35,20 +43,26 @@ router.post("/", async (req, res) => {
   try {
     const savedPost = await post.save();
     res.status(201).send(savedPost);
-    console.log(req.user);
   } catch (err) {
     console.log(err);
+    res
+      .status(500)
+      .status({ message: "Coult not create post. Please try again" });
   }
 });
 
 //Update a post
-router.patch("/:postId", async (req, res) => {
+router.patch("/:postId", getUser, auth, async (req, res) => {
   const { gameTitle, genre, description } = req.body;
+  const { userName } = req.user;
   const postId = req.params.postId;
-
   try {
     const post = await Post.findById(postId);
     if (!post) return res.status(404).send({ message: "Post not found" });
+    if (userName !== post.userName)
+      return res
+        .status(401)
+        .send({ message: "You can only update your own posts" });
     await Post.updateOne(
       { _id: postId },
       {
@@ -66,28 +80,37 @@ router.patch("/:postId", async (req, res) => {
 });
 
 //Mark game as exchanged
-router.patch("/exchange/:postId", async (req, res) => {
+router.patch("/exchange/:postId", getUser, auth, async (req, res) => {
   const postId = req.params.postId;
+  const { userName } = req.user;
   try {
     let post = await Post.findById(postId);
     if (!post) return res.status(404).send({ message: "Post not found" });
+    if (userName !== post.userName)
+      return res.status(401).send({
+        message: "You can only mark or unmark posts that you created",
+      });
     const updateBoolean = post.exchanged ? false : true;
     await Post.updateOne({ _id: postId }, { exchanged: updateBoolean });
     post = await Post.findById(postId);
     res.status(200).send(post);
   } catch (err) {
     console.log(err);
-    res.status(500).send({ error: "Something went wrong" });
+    res.status(500).send({ error: "Could not update post. Please try again" });
   }
 });
 
 //Delete a post
-router.delete("/:postId", async (req, res) => {
+router.delete("/:postId", getUser, auth, async (req, res) => {
+  const { userName } = req.user;
   try {
     const post = await Post.findById(req.params.postId);
 
     if (!post) return res.status(404).send({ message: "Post not found" });
-
+    if (userName !== post.userName)
+      return res
+        .status(401)
+        .send({ message: "You can only delete your own posts" });
     await Post.deleteOne({ _id: req.params.postId });
     res.send("Your post has been succesfully deleted");
   } catch (err) {
